requirements.txt
# ===== requirements.txt (Updated) =====

# Core framework
streamlit>=1.39.0

# Local vector processing
langchain>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
faiss-cpu>=1.8.0

# PDF and image processing
PyMuPDF>=1.24.0
Pillow>=10.4.0

# OCR models
transformers>=4.45.0
torch>=2.1.0
torchvision>=0.16.0
easyocr>=1.7.0

# Local LLM support
langchain-ollama>=0.3.0

# Utilities
python-dotenv>=1.0.1
numpy>=1.24.0

# Optional: For better performance
accelerate>=0.34.0
sentence-transformers>=3.0.0


